Split Aethel views; rename types
