+++ backend/aethel_db/urls.py
@@ -1,7 +1,8 @@
 from django.urls import path
-from aethel_db.views import AethelQueryView, AethelDetailView
+from aethel_db.views.AethelDetailView import AethelDetailView
+from aethel_db.views.AethelListView import AethelListView
 
 urlpatterns = [
-    path("", AethelQueryView.as_view(), name="aethel-list"),
+    path("", AethelListView.as_view(), name="aethel-list"),
     path("sample", AethelDetailView.as_view(), name="aethel-detail"),
 ]
