+++ backend/aethel_db/views/AethelListView.py
@@ -1,25 +1,15 @@
 from dataclasses import asdict, dataclass, field
-from enum import Enum
-from typing import List, Optional
 
 from django.http import HttpRequest, JsonResponse
 from rest_framework import status
 from rest_framework.views import APIView
-from spindle.utils import serialize_phrases_with_infix_notation
-from aethel_db.search import (
-    match_type_with_phrase,
-    match_word_with_phrase,
-)
-from aethel.frontend import Sample
-from .models import dataset
 
-
-def aethel_status():
-    return dataset is not None
+from aethel_db.search import match_type_with_phrase, match_word_with_phrase
+from aethel_db.models import dataset
 
 
 @dataclass
-class AethelSamplePhrase:
+class AethelListSamplePhrase:
     display: str
     highlight: bool
 
@@ -27,7 +17,7 @@ class AethelSamplePhrase:
 @dataclass
 class AethelListSample:
     name: str
-    phrases: List[AethelSamplePhrase] = field(default_factory=list)
+    phrases: list[AethelListSamplePhrase] = field(default_factory=list)
 
 
 @dataclass
@@ -35,7 +25,7 @@ class AethelListItem:
     lemma: str
     word: str
     type: str
-    samples: List[AethelListSample] = field(default_factory=list)
+    samples: list[AethelListSample] = field(default_factory=list)
 
     def serialize(self):
         out = asdict(self)
@@ -52,7 +42,7 @@ class AethelListResponse:
     """
 
     results: dict[tuple[str, str, str], AethelListItem] = field(default_factory=dict)
-    error: Optional[str] = None
+    error: str | None = None
 
     def get_or_create_result(self, lemma: str, word: str, type: str) -> AethelListItem:
         """
@@ -74,7 +64,7 @@ class AethelListResponse:
         )
 
 
-class AethelQueryView(APIView):
+class AethelListView(APIView):
     def get(self, request: HttpRequest) -> JsonResponse:
         word_input = self.request.query_params.get("word", None)
         type_input = self.request.query_params.get("type", None)
@@ -111,7 +101,7 @@ class AethelQueryView(APIView):
                     new_sample = AethelListSample(name=sample.name, phrases=[])
                     for index, sample_phrase in enumerate(sample.lexical_phrases):
                         highlighted = index == phrase_index
-                        new_phrase = AethelSamplePhrase(
+                        new_phrase = AethelListSamplePhrase(
                             display=sample_phrase.string,
                             highlight=highlighted,
                         )
@@ -119,80 +109,3 @@ class AethelQueryView(APIView):
                     result.samples.append(new_sample)
 
         return response_object.json_response()
-
-
-class AethelDetailError(Enum):
-    NO_QUERY_INPUT = "NO_QUERY_INPUT"
-    SAMPLE_NOT_FOUND = "SAMPLE_NOT_FOUND"
-    MULTIPLE_FOUND = "MULTIPLE_FOUND"
-
-
-aethel_detail_status_codes = {
-    AethelDetailError.NO_QUERY_INPUT: status.HTTP_400_BAD_REQUEST,
-    AethelDetailError.SAMPLE_NOT_FOUND: status.HTTP_404_NOT_FOUND,
-    AethelDetailError.MULTIPLE_FOUND: status.HTTP_500_INTERNAL_SERVER_ERROR,
-}
-
-
-@dataclass
-class AethelDetailResult:
-    sentence: str
-    name: str
-    term: str
-    subset: str
-    phrases: list[dict]
-
-
-@dataclass
-class AethelDetailResponse:
-    result: Optional[AethelDetailResult] = None
-    error: Optional[AethelDetailError] = None
-
-    def parse_sample(self, sample: Sample) -> None:
-        self.result = AethelDetailResult(
-            sentence=sample.sentence,
-            name=sample.name,
-            term=str(sample.proof.term),
-            subset=sample.subset,
-            phrases=serialize_phrases_with_infix_notation(sample.lexical_phrases),
-        )
-
-    def json_response(self) -> JsonResponse:
-        result = asdict(self.result) if self.result else None
-        status_code = (
-            aethel_detail_status_codes[self.error] if self.error else status.HTTP_200_OK
-        )
-
-        return JsonResponse(
-            {
-                "result": result,
-                "error": self.error,
-            },
-            status=status_code,
-        )
-
-
-class AethelDetailView(APIView):
-    def get(self, request: HttpRequest) -> JsonResponse:
-        query_input = self.request.query_params.get("sample-name", None)
-
-        if query_input is None:
-            response = AethelDetailResponse(error=AethelDetailError.NO_QUERY_INPUT)
-            return response.json_response()
-
-        samples = dataset.find_by_name(query_input)
-
-        if len(samples) == 0:
-            response = AethelDetailResponse(error=AethelDetailError.SAMPLE_NOT_FOUND)
-            return response.json_response()
-
-        if len(samples) > 1:
-            response = AethelDetailResponse(error=AethelDetailError.MULTIPLE_FOUND)
-            return response.json_response()
-
-        sample = samples[0]
-
-        response = AethelDetailResponse()
-        response.parse_sample(sample)
-
-        return response.json_response()
