+++ backend/aethel_db/views/aethel_status.py
@@ -1,0 +1,5 @@
+from aethel_db.models import dataset
+
+
+def aethel_status():
+    return dataset is not None
