+++ backend/parseport/views.py
@@ -1,7 +1,7 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 
-from aethel_db.views import aethel_status
+from aethel_db.views.aethel_status import aethel_status
 from spindle.views import spindle_status
